scripts/mp4converter/pyproject.toml
[project]
name = "mp4converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mango", email = "imango@outlook.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "tqdm>=4.67.1",
]

[project.scripts]
mp4-converter = "mp4converter:main"

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"
